20230919-lsm-01 commit

+++ src/main/java/com/daalgae/daalgaeproject/board/service/BoardService.java
@@ -7,6 +7,7 @@ import com.daalgae.daalgaeproject.common.exception.board.ReplyRegistException;
 import com.daalgae.daalgaeproject.common.exception.board.ReplyRemoveException;
 import com.daalgae.daalgaeproject.common.exception.thumbnail.ThumbnailRegistException;
 import com.daalgae.daalgaeproject.common.paging.SelectCriteria;
+import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 import java.util.Map;
@@ -20,10 +21,10 @@ public interface BoardService {
     public List<BoardDTO> selectBoardList(SelectCriteria selectCriteria);
 
     /* 게시글 상세 페이지 조회용 메소드 */
-    public BoardDTO selectBoardDetail(Long no);
+    public BoardDTO selectBoardDetail(int no);
 
     /* 해당 게시글의 전체 댓글 조회용 메소드 */
-    public List<ReplyDTO> selectAllReplyList(Long boardNo);
+    public List<ReplyDTO> selectAllReplyList(int boardNo);
 
 
     /* 댓글 등록용 메소드 */
@@ -43,5 +44,5 @@ public interface BoardService {
 
 
     /* 게시글 상세 페이지 조회용 메소드 */
-    public BoardDTO selectThumbnailDetail(Long no) ;
+    public BoardDTO selectThumbnailDetail(int no) ;
 }
